simple Image clean up
